refactor(plan_SERVERSPLIT slice 1): the cookies-storage bodies leave server.py
The first three tool bodies move out of the god file, into
embedded/tool_sections/cookies_storage.py. cookies-storage is the smallest
section that still reaches for the whole shared-dependency set — browser_manager,
network_interceptor, _with_cdp_timeout, _require_tab and ToolError — so it proves
the slice-0 mechanism on three tools before it is trusted with four hundred lines.

A pure move. The bodies are verbatim; the only edits are the dropped registration
decorator (contract rule 2: registration is driven from server.py's binding loop,
once per execution of that module body, so the runpy __main__ load still gets a
full app) and the rewrite of the singleton reads to rt.<name>, resolved against
tool_runtime at CALL time so conftest's patched_server still reaches them
(contract rule 3). Docstrings and signatures are byte-identical, and
tests/goldens/tool_surface.json — the HARD golden — is unchanged.

Verified: 2334 passed / 1 skipped on the unit lane; tool surface IDENTICAL to the
golden; SECTION_TOOLS 94 with per-app [94, 94, 94] across the three loaded
identities; --list-sections subprocess "Total: 94 tools"; gen_release_contract
--check clean; ruff format/check, vulture, ty and the suppression-owner gate
clean. The real-stdio transport lane (test_e2e_transport_cookies.py, real
headless Chrome: set_cookie -> get_cookies -> assert the value -> clear_cookies)
passes through the moved bodies, so plan §7 R7's known .fn-seam quirk is
confirmed absent over the wire as well as on the seam.

server.py: 3342 -> 3248 LOC, cap ratcheted DOWN to the measured actual.
tests/source_scan.py's floor ratchets UP 2 -> 3 in the same commit: waiting for
slice 11 to raise it in one jump would leave ten slices during which a dropped
section module reads as a legitimately smaller set rather than a collapsed one.
