+++ CLAUDE.md
@@ -56,7 +56,7 @@ Package root: `src/stealth_chrome_devtools_mcp/`. Two console scripts (`pyprojec
 **Lifecycle & transport**
 | File | Owns |
 |---|---|
-| `server.py` | the real MCP server — all 94 tool bodies + `app_lifespan` + `mcp`/`registry` + the **binding loop** that drives registration from THIS file's module body, once per execution of it (so the canonical import, the bare-name spec load and the runpy `__main__` load each get a full 94-tool app — a section module that decorated itself would register into the first execution only). The shared substrate the bodies reach for lives in `tool_runtime.py` |
+| `server.py` | the real MCP server — the **91 tool bodies not yet moved** (the 3 `cookies-storage` bodies live in `tool_sections/cookies_storage.py` as of plan_SERVERSPLIT slice 1) + `app_lifespan` + `mcp`/`registry` + the **binding loop** that drives registration from THIS file's module body, once per execution of it (so the canonical import, the bare-name spec load and the runpy `__main__` load each get a full 94-tool app — a section module that decorated itself would register into the first execution only). The shared substrate the bodies reach for lives in `tool_runtime.py` |
 | `singleton.py` | **backend lifecycle + the stdio proxy** — liveness probes (`_backend_http_ready`, `_probe_backend_status`), port selection (`_select_backend_port`, `DEFAULT_PORT`), the one identity+readiness reuse gate (`_same_identity_backend_ready`, `_source_fingerprint`, `REUSE_PATIENCE_SECONDS` — which spends its patience through `scheduling_lag.FairWindow`, F-856), cold-start lock (`_start_backend_holding_lock`), `run_stdio_proxy`. The `server.json` record moved to `backend_registry.py` (path names re-exported here for legacy callers); the proxy's *recovery from a dead backend* moved to `proxy_selfheal.py`; the watchdog LOOP moved to `backend_watchdog.py` — `_watch_backend_liveness` stays here as the wiring that knows which probes are ours |
 | `backend_watchdog.py` | **THE one home for the proxy's mid-session liveness watchdog** — the SLOW witness (`watch_liveness`): F-820's strikes plus the confirmation phase. A leaf: both probes arrive as arguments, so it never imports `singleton` and the dead-vs-busy policy stays single-homed in the reuse gate |
 | `scheduling_lag.py` | **THE one home for "was this process scheduled fairly, and what does a time budget owe it when it was not"** (F-856) — `FairWindow`, whose budget is charged in fair seconds (elapsed ÷ the lag its own naps measured), plus `MAX_STRETCH`, `REPORT_FACTOR` and the `proxy: patience extended under starvation` lifecycle report. It never decides alive-or-dead: only "has this window been spent", so `proxy_selfheal`'s ONE heal path is untouched. A leaf; the `_now`/`_wait` module functions are its single timing seam |
@@ -67,8 +67,8 @@ Package root: `src/stealth_chrome_devtools_mcp/`. Two console scripts (`pyprojec
 | `desktop_launch.py` | **THE one home for "hand a launch to the logged-on user's desktop"** (F-810, Windows only) — the availability probe (`available`, `_active_console_session_id`), the two predicates the guard and the spawn pipeline consult (`can_deliver_headed_window`, `should_delegate`), the Task Scheduler round trip (`launch_and_attach`, `_schtasks`) and the attached-browser process shim (`pid_shim`). The OS places the process; the tool still never picks a session |
 | `browser_pid_registry.py` | **THE one home for `browser_pids.json`** — its schema, the owner stamp on every entry (`owner_pid`, `owner_create_time`), and the read-merge-write protocol every writer shares |
 | `tool_registry.py` | `SECTION_TOOLS` + `ToolRegistry.section_tool` (registration, section gating, correlation-id stamping, the surrogate-safe return wrap) + the canonical **verb taxonomy** (module docstring) |
-| `tool_runtime.py` | **THE one home for what a tool body reaches for beyond its own arguments** (plan_SERVERSPLIT slice 0) — the four constructed singletons (`browser_manager`, `network_interceptor`, `dom_handler`, `cdp_function_executor`), the re-exported module singletons, the four tuned knobs (`CDP_OPERATION_TIMEOUT`, `MAX_TIMEOUT_MS`, `EXECUTE_SCRIPT_TIMEOUT`, `MAX_USER_SCRIPT_BYTES`) and the guards that enforce them (`_BLOCKING_SCRIPT_PATTERNS`, `_script_rejection_reason`, `_clamp_timeout`, `_with_cdp_timeout`). A leaf — it imports no `mcp`, no `registry`, no `server`. Being a normal module it is loaded ONCE, unlike `server.py`, so a body that resolves `rt.<name>` at call time has exactly ONE patchable home (`tests/conftest.py`'s `patched_server`); `__all__` states that surface. **Slice-0 truth: all 94 tool bodies are still in `server.py`** and reach these through a migration alias import there, deleted in slice 12 |
-| `tool_sections/` | the subpackage the 94 tool bodies move into, one module per `SECTION_TOOLS` section. `__init__.py` owns the contract (a section module exports only `SECTION` + `TOOLS`, **never** decorates, **never** imports `mcp`/`registry`/`server`, resolves everything as `rt.<name>` at call time) and `SECTION_MODULES`, the one enumeration `server.py`'s binding loop and `tests/source_scan.py` both walk. **Slice-0 truth: `SECTION_MODULES` is `()`** — the mechanism ships before any tool moves |
+| `tool_runtime.py` | **THE one home for what a tool body reaches for beyond its own arguments** (plan_SERVERSPLIT slice 0) — the four constructed singletons (`browser_manager`, `network_interceptor`, `dom_handler`, `cdp_function_executor`), the re-exported module singletons, the four tuned knobs (`CDP_OPERATION_TIMEOUT`, `MAX_TIMEOUT_MS`, `EXECUTE_SCRIPT_TIMEOUT`, `MAX_USER_SCRIPT_BYTES`) and the guards that enforce them (`_BLOCKING_SCRIPT_PATTERNS`, `_script_rejection_reason`, `_clamp_timeout`, `_with_cdp_timeout`). A leaf — it imports no `mcp`, no `registry`, no `server`. Being a normal module it is loaded ONCE, unlike `server.py`, so a body that resolves `rt.<name>` at call time has exactly ONE patchable home (`tests/conftest.py`'s `patched_server`); `__all__` states that surface. **Migration truth: the bodies still in `server.py` reach these as bare names** through a migration alias import there, deleted in slice 12; a body that has moved into `tool_sections/` reaches them as `rt.<name>` |
+| `tool_sections/` | the subpackage the 94 tool bodies move into, one module per `SECTION_TOOLS` section. `__init__.py` owns the contract (a section module exports only `SECTION` + `TOOLS`, **never** decorates, **never** imports `mcp`/`registry`/`server`, resolves everything as `rt.<name>` at call time) and `SECTION_MODULES`, the one enumeration `server.py`'s binding loop and `tests/source_scan.py` both walk. **Migration truth after slice 1: `cookies_storage.py` (3 tools).** The other ten sections are still in `server.py` and arrive one slice at a time; `tests/source_scan.py`'s floor ratchets UP by one with each, so a dropped module reads as a collapsed set rather than a smaller one |
 | `tool_errors.py` | the error convention — `ToolError`, `InstanceNotFoundError`, `_require_tab`, `_require_browser` |
 | `logging_setup.py` | the observability spine — `resolve_log_dir`, `configure_logging`, `with_correlation_id`, `CorrelationIdFilter`; `with_correlation_id` is also **THE one place a failed tool call is recorded** (`_record_tool_failure` → `debug_logger.log_tool_failure`, F-835) — never add a per-tool `except` that logs; **the one home for log-file retention** — `roll_boot_log` (the launcher-side boot-log rotation, F-830) and `prune_old_logs` + its dead-backend post-mortem exemption (F-840); plus `backend_uvicorn_config` — the one home for the backend's uvicorn run-config (access logging off, graceful-shutdown timeout) |
 | `process_cleanup.py` | orphan reaping — side-effect-free `__init__`, `activate()` at serve boundary (handlers armed synchronously; the REAP is handed to `serve_startup.after_serving`, F-856), `recover_orphans()` seam |
