+++ src/stealth_chrome_devtools_mcp/embedded/server.py
@@ -1360,100 +1360,6 @@ async def modify_headers(instance_id: str, headers: dict[str, str]) -> bool:
     )
 
 
-@section_tool("cookies-storage")
-async def get_cookies(
-    instance_id: str, urls: list[str] | None = None
-) -> list[dict[str, Any]]:
-    """
-    Get cookies for current page or specific URLs.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        urls (Optional[List[str]]): Optional list of URLs to get cookies for.
-
-    Returns:
-        List[Dict[str, Any]]: List of cookies.
-    """
-    tab = await _require_tab(browser_manager, instance_id)
-    return await _with_cdp_timeout(
-        network_interceptor.get_cookies(tab, urls), instance_id=instance_id
-    )
-
-
-@section_tool("cookies-storage")
-async def set_cookie(
-    instance_id: str,
-    name: str,
-    value: str,
-    url: str | None = None,
-    domain: str | None = None,
-    path: str = "/",
-    secure: bool = False,
-    http_only: bool = False,
-    same_site: str | None = None,
-) -> bool:
-    """
-    Set a cookie.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        name (str): Cookie name.
-        value (str): Cookie value.
-        url (Optional[str]): The request-URI to associate with the cookie.
-        domain (Optional[str]): Cookie domain.
-        path (str): Cookie path.
-        secure (bool): Secure flag.
-        http_only (bool): HttpOnly flag.
-        same_site (Optional[str]): SameSite — 'Strict', 'Lax' or 'None' (any case).
-
-    Returns:
-        bool: True if set successfully.
-    """
-    tab = await _require_tab(browser_manager, instance_id)
-
-    if not url and not domain:
-        current_url = tab.url if hasattr(tab, "url") else None
-        if current_url:
-            url = current_url
-        else:
-            raise ToolError("At least one of 'url' or 'domain' must be specified")
-
-    cookie = {
-        "name": name,
-        "value": value,
-        "path": path,
-        "secure": secure,
-        "http_only": http_only,
-    }
-    if url:
-        cookie["url"] = url
-    if domain:
-        cookie["domain"] = domain
-    if same_site:
-        cookie["same_site"] = same_site
-    return await _with_cdp_timeout(
-        network_interceptor.set_cookie(tab, cookie), instance_id=instance_id
-    )
-
-
-@section_tool("cookies-storage")
-async def clear_cookies(instance_id: str, url: str | None = None) -> bool:
-    """
-    Clear cookies.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        url (Optional[str]): Optional URL to clear cookies for (clears all if not specified).
-
-    Returns:
-        bool: True if cleared successfully.
-    """
-    tab = await _require_tab(browser_manager, instance_id)
-    return await _with_cdp_timeout(
-        network_interceptor.clear_cookies(tab, url), instance_id=instance_id
-    )
-
-
 @mcp.resource("browser://{instance_id}/state")
 async def get_browser_state_resource(instance_id: str) -> str:
     """
